Use constant 'from' address.
Spam filters will reject a message because the server sending the message
is not the official mail server for that user (i.e. SPF records). Although
it is nice to have the "From" being the e-mail address of the form people
are filling out it just doesn't work well.

To resolve this we are simply setting the from the same as the 'to'.